refactor(lib): address review — narrow the names, move the generic helper
Two naming points from review:

  - `explainDataError` claimed more than it does, so it is now
    `explainCodecError` (and `TDataErrorExplanation` follows it to
    `TCodecErrorExplanation`).
  - `flattenErrorMessage` is not codec-specific at all — it moves to
    `src/utils/errorHandling.ts` next to `getErrorMessage` and
    `toNormalizedError`, with its unit tests alongside it in
    `tests/unit/utils/errorHandling.test.ts`.

No behaviour change.
